Add TCP/IP example
