+++ Modelica_DeviceDrivers/Blocks/Examples/package.order
@@ -13,3 +13,4 @@ TestInputKeyboard
 TestInputKeyboardKey
 TestHardwareIOComedi
 TestSerialPackager_UDPMinimalForOMCTest
+TestSerialPackager_TCPIP
